WebStorage/tests/approved/event_local_storagearea.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>event_local_storageArea</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		var t = async_test("storageArea property test of local event");

		function onStorageEvent(event) {
			if (event.storageArea != null) {
				try {
					for (var i = 0; i < event.storageArea.length; i++) {
						var key = event.storageArea.key(i);
						var value = event.storageArea.getItem(key);
						var msg = "key: " + key + "; value: " + value;

						t.step(function() {
						   assert_equals(msg, "key: name; value: user1");
						  });
						t.done();
					}
				} catch (e) {
					t.step(function() {
						assert_unreached("Storage Interface error!");
					});
					t.done();
				}
			}
		}

		if (window.addEventListener) {
            window.addEventListener('storage', onStorageEvent, false);
        }

        else {
			test(function() {
				assert_unreached("DOM event subscribtion is not supported.");
			}, "existence of event listener")
        }

		var el = document.createElement("iframe");
		el.setAttribute('id', 'ifrm');
		el.setAttribute('src', 'iframe/local_set_item_iframe.html');
		el.style.border='0px';
		el.style.width='0px';
		el.style.height='0px'		
		document.body.appendChild(el);

	</script>

 </body>
</html>
